require dirs configuration to be a list not a str
I encountered an issue where I wasn't getting the behavior I expected
out of a certain configuration.  It turned out that I had some code
which was treating a string like a list and python was perfectly happy
to iterate over the characters in the string.  I did a bit of research
and it is actually fairly difficult to tell the difference between a
python string and a list in a way which behaves how you would expect and
doesn't violate duck-typing.  In the end I chose to violoate duck-typing
and check that the parameter isn't a str or unicode.  I'm not thrilled
with it, but I'd rather get an explicit error message if I mess up my
configuration again.